Many changes - Read detailed description.
Not currently running, but there are many changes in this update and its
a good time to mark a milestone.

The Code inheritance heirarchy is being refactored into a monolithic
Code class, removing all of the subclasses.

The only purpose of the subclasses was to define an execute() method for
each language. This can be done much more efficiently by simply calling
the appropriate execute message for each language within the Code class'
execute method. Additionally, this design is much more practical for
QML, as QML seems to require the redefinition of all parent class properties
within subclasses.
This will also lighten up the application's file count, significantly reduce
the lines of code and remove the need for a code_factory.

The languages file has been refactored into a Languages static method
class. This allows easy support for QML while retaining the command line
interface support.

QML
This commit adds significant support for a QML user interface. While not
currently operable, QML functionality will be available shortly.